shared process - move the shared process back to IPC folder
We still have to figure out where to best put these files
