+++ src/vs/platform/ipc/electron-browser/sharedProcessService.ts
