+++ src/vs/platform/ipc/electron-main/sharedProcessMainService.ts
