+++ src/vs/platform/ipc/node/sharedProcess.ts
